--- packages/plugin-misc/src/crossmint/index.ts
@@ -1,2 +1,0 @@
-export * from "./tools/checkout";
-export * from "./tools/confirm-order";
